Move doc extensions into docs/_ext folder
